+++ flags/flag_up_x.c
@@ -6,13 +6,13 @@
 /*   By: ebini <ebini@student.42lyon.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/03 23:24:07 by ebini             #+#    #+#             */
-/*   Updated: 2024/12/22 15:50:26 by ebini            ###   ########lyon.fr   */
+/*   Updated: 2024/12/22 16:42:48 by ebini            ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "utils.h"
 
-int	print_up_x(va_list value)
+size_t	print_up_x(va_list value)
 {
 	return (print_base(va_arg(value, unsigned), "0123456789ABCDEF", 16));
 }
